requirements.txt
openai
requests
pandas
openpyxl
python-dotenv
rapidfuzz
supabase
